+++ src/flickr_photos_api/api/__init__.py
@@ -268,76 +268,6 @@ class FlickrPhotosApi(BaseApi):
             "location": location,
         }
 
-    def get_public_photos_by_user(
-        self, user_url: str, page: int = 1, per_page: int = 10
-    ) -> CollectionOfPhotos:
-        """
-        Get all the public photos by a user on Flickr.
-        """
-        user = self.lookup_user_by_url(url=user_url)
-
-        # See https://www.flickr.com/services/api/flickr.people.getPublicPhotos.html
-        resp = self._get_page_of_photos(
-            method="flickr.people.getPublicPhotos",
-            params={
-                "user_id": user["id"],
-                "extras": ",".join(self.extras),
-            },
-            page=page,
-            per_page=per_page,
-        )
-
-        return {
-            "total_photos": resp["total_photos"],
-            "page_count": resp["page_count"],
-            "photos": [self._to_photo(photo_elem) for photo_elem in resp["elements"]],
-        }
-
-    def lookup_group_from_url(self, *, url: str) -> GroupInfo:
-        """
-        Given the link to a group's photos or profile, return some info.
-        """
-        resp = self.call(method="flickr.urls.lookupGroup", params={"url": url})
-
-        # The lookupUser response is of the form:
-        #
-        #       <group id="34427469792@N01">
-        #         <groupname>FlickrCentral</groupname>
-        #       </group>
-        #
-        group_elem = find_required_elem(resp, path=".//group")
-
-        return {
-            "id": group_elem.attrib["id"],
-            "name": find_required_text(group_elem, path="groupname"),
-        }
-
-    def get_photos_in_group_pool(
-        self, group_url: str, page: int = 1, per_page: int = 10
-    ) -> PhotosInGroup:
-        """
-        Get all the photos in a group pool.
-        """
-        group_info = self.lookup_group_from_url(url=group_url)
-
-        # See https://www.flickr.com/services/api/flickr.groups.pools.getPhotos.html
-        resp = self._get_page_of_photos(
-            method="flickr.groups.pools.getPhotos",
-            params={
-                "group_id": group_info["id"],
-                "extras": ",".join(self.extras),
-            },
-            page=page,
-            per_page=per_page,
-        )
-
-        return {
-            "photos": [self._to_photo(photo_elem) for photo_elem in resp["elements"]],
-            "page_count": resp["page_count"],
-            "total_photos": resp["total_photos"],
-            "group": group_info,
-        }
-
     def get_photos_with_tag(
         self, tag: str, page: int = 1, per_page: int = 10
     ) -> CollectionOfPhotos:
